requirements.txt
# Stock Trading Bot Dependencies
# Python 3.10+ required

# Data & Analysis
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0  # Technical Analysis library

# Broker Integration
alpaca-py>=0.21.0

# Database
sqlalchemy>=2.0.0

# Configuration
pyyaml>=6.0.0
python-dotenv>=1.0.0

# HTTP & API
requests>=2.31.0
aiohttp>=3.9.0

# Dashboard
streamlit>=1.31.0
plotly>=5.18.0

# Notifications
discord-webhook>=1.3.0

# Scheduling
schedule>=1.2.0
pytz>=2024.1

# ML (optional, for signal layer)
scikit-learn>=1.4.0
joblib>=1.3.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Logging
python-json-logger>=2.0.0

# Linting
ruff>=0.3.0

# Type checking (dev)
mypy>=1.8.0
pandas-stubs>=2.0.0
